fix(swift): мелочи из ревью — тип кода ошибки, Sendable, протухшие докблоки
.unrecognized хранил Int32, тогда как ChupaErrorCode.rawValue — UInt32.
Между ними стоял Int32(_:), который на непомещающемся значении роняет процесс.
Случай существует ровно затем, чтобы новый код ошибки от обновлённого движка
не уронил хост, — и оказался бы единственным, который не выживает. Значений
таких не будет никогда (кодов восемь, растут по единице), так что это не риск
падения, а неправильно выбранный тип. Теперь UInt32, сужения нет.

Error помечен Sendable: все его поля сендабельны, а ErrorCode был помечен ещё
раньше. Без этого хост под строгим параллелизмом не пробросил бы ошибку из
фоновой очереди.

Докблоки Context, Expression и Script утверждали, что единица компиляции может
пережить контекст «в любом порядке». Это контракт C API, а не свойство обвязки:
каждая единица держит на контекст сильную ссылку, поэтому контекст всегда
переживает свои выражения и скрипты. Обвязка строго безопаснее C API, и сказано
это теперь как своё свойство.

Комментарий в Aliases.swift ссылался на CSEvaluation, удалённый при упрощении.

14 тестов зелёные.

+++ Sources/ChupaScript/ErrorCode.swift
@@ -50,7 +50,12 @@ public enum ErrorCode: Sendable, Equatable {
     /// Случай нужен, чтобы обновление движка не превращалось в аварию на
     /// стороне хоста: неизвестный код доедет до него числом, а не потеряется
     /// и не свалит разбор.
-    case unrecognized(Int32)
+    ///
+    /// Тип совпадает с `ChupaErrorCode.rawValue` — сужать до `Int32` нельзя:
+    /// `Int32(_:)` на непомещающемся значении роняет процесс, и ветка,
+    /// существующая ради выживания, оказалась бы единственной, которая не
+    /// выживает.
+    case unrecognized(UInt32)
 }
 
 extension ErrorCode {
@@ -66,7 +71,7 @@ extension ErrorCode {
         case CHUPA_ERR_DATA:   self = .data
         case CHUPA_ERR_USAGE:  self = .usage
         case CHUPA_ERR_MEMORY: self = .memory
-        default:               self = .unrecognized(Int32(raw.rawValue))
+        default:               self = .unrecognized(raw.rawValue)
         }
     }
 }
